+++ backend/src/modules/app-service/app-service.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { AppServiceController } from '@modules/app-service/app-service.controller';
+import { AppServiceController } from './controllers/app-service.controller';
 
 @Module({
   controllers: [AppServiceController],
