+++ backend/src/modules/app-service/controllers/app-service.controller.ts
@@ -1,5 +1,5 @@
+import { Public } from '@modules/auth/services/auth.decorator';
 import { Controller, Get, VERSION_NEUTRAL, Version } from '@nestjs/common';
-import { Public } from '@modules/auth/auth.decorator';
 
 @Controller({
     version: VERSION_NEUTRAL,
