+++ backend/src/modules/auth/services/auth.guard.ts
@@ -7,8 +7,8 @@ import {
 import { JwtService } from '@nestjs/jwt';
 import { jwtConstants } from './constants';
 import { Request } from 'express';
-import { IS_PUBLIC_KEY } from '@modules/auth/auth.decorator';
 import { Reflector } from '@nestjs/core';
+import { IS_PUBLIC_KEY } from './auth.decorator';
 @Injectable()
 export class AuthGuard implements CanActivate {
     constructor(private jwtService: JwtService, private reflector: Reflector) { }
