+++ backend/src/modules/user/controllers/users.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UsersController } from './users.controller';
+
+describe('UsersController', () => {
+    let controller: UsersController;
+
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            controllers: [UsersController],
+        }).compile();
+
+        controller = module.get<UsersController>(UsersController);
+    });
+
+    it('should be defined', () => {
+        expect(controller).toBeDefined();
+    });
+});
