+++ backend/src/modules/user/services/users.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, NotImplementedException } from '@nestjs/common';
-import { User } from './User';
+import { User } from '../models/User';
 
 export const mockUsersCollection = [
     new User({ password: '123', username: 'one' }),
