+++ backend/test/jest-e2e.json
@@ -13,6 +13,7 @@
     "preset": "ts-jest",
     "moduleDirectories": ["node_modules", "<rootDir>"],
     "moduleNameMapper": {
-        "@modules/(.*)": "<rootDir>/../src/modules/$1"
+        "@modules/(.*)": "<rootDir>/../src/modules/$1",
+        "@src/(.*)": "<rootDir>../src/$1"
     }
 }
